Compiled both keys into a single class. No reason not to
